[Nexters/covid-letter-web#83] :lipstick: 마크업 구조 수(컴포넌트 분리 등)
